% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{GetPrimes}
\alias{GetPrimes}
\title{Compute N Primes.}
\usage{
GetPrimes(n)
}
\arguments{
\item{n}{number of primes.}
}
\value{
sequence of primes.
}
\description{
This function computes the first n primes.
}
