% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_combos.R
\name{PersonVars}
\alias{PersonVars}
\title{Computes the abundance of each 'variant' for each person}
\usage{
PersonVars(person_taxa, taxa_var)
}
\arguments{
\item{person_taxa}{matrix of person vs taxa abundance.}

\item{taxa_var}{matrix of taxa vs variant (presence T/F)}
}
\value{
sparse representation of abundance of each variant for each person.
}
\description{
Computes the abundance of each 'variant' for each person
}
