% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{SetTime}
\alias{SetTime}
\title{Recomputes the time.}
\usage{
SetTime(t = NA)
}
\arguments{
\item{t}{current time.}
}
\value{
new time.
}
\description{
This function prints the time elapsed between now and the last time SetTime was called.
}
