% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_combos.R
\name{TaxaVariantCombos}
\alias{TaxaVariantCombos}
\title{Computes combinations of n_df variants.}
\usage{
TaxaVariantCombos(taxa_variant_sparse, i_taxa, n_df, index_lookup)
}
\arguments{
\item{taxa_variant_sparse}{sparse matrix of taxa vs variants.}

\item{i_taxa}{which taxa to compute variants for.}

\item{n_df}{number of variants (combos of 2, 3, 4, etc.) to consider in each combination.}
}
\value{
sparse representation of taxa variant matrix.
}
\description{
Computes combinations of n_df variants.
}
