% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_combos.R
\name{TaxaVariantSparse}
\alias{TaxaVariantSparse}
\title{Turns matrix of taxa vs variants into a sparse matrix.}
\usage{
TaxaVariantSparse(taxa_vars, variant_vals)
}
\arguments{
\item{taxa_vars}{matrix of taxa vs variants.}

\item{variant_vals}{list of variant values to check for.}
}
\value{
sparse representation of taxa variant matrix.
}
\description{
Turns matrix of taxa vs variants into a sparse matrix.
}
