% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_combos.R
\name{TrimTaxaVar}
\alias{TrimTaxaVar}
\title{Computes the abundance of each 'variant' for each person}
\usage{
TrimTaxaVar(taxa_var, var_names, has_vals = F)
}
\arguments{
\item{taxa_var}{matrix of person vs taxa abundance.}

\item{has_vals}{whether or not the sparse representation contains values (T) or if it just contains 0/1s (F).}
}
\value{
trimmed
}
\description{
This function trims the a matrix to only include non-duplicated columns.
}
