% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{WilcoxonRankFastPaired}
\alias{WilcoxonRankFastPaired}
\title{Computes the p-value of the wilcoxon rank test.}
\usage{
WilcoxonRankFastPaired(ranks, signs)
}
\arguments{
\item{ranks}{list of rank differences.}

\item{signs}{list sof sign differences.}
}
\value{
p-value.
}
\description{
This function performs a wilcox rank sum test given the ranking and the signs.
}
